Latest version: 04/06/2015
This is the latest version for our project.
Now the project is much smarter.  All of the mf struct and algorithms
are generated from our source code.
This version can only implement a input for one grouping variable and
also no "HAVING condition" included in this version.
The code is a little bit complicated.  Please make sure fully
understanding the code before we can move to next step.